+++ CHANGELOG.md
@@ -11,10 +11,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Custom Metadata Report now supports looping items in a Catalogue (use `$foreach CatalogueItem` to start and `$end` to end)
+- Added help to 'New Project' user interface
 
 ### Changed
 
-- Improved usability of selecting multiple datasets in the New Project user interface
+- Improved usability of selecting multiple datasets in the 'New Project' user interface
 - When in multiple selection mode, double clicking a row in the object selection dialog will add it to the selection (previously would close the dialog with the double clicked item as the sole selected item)
 
 ## [4.1.3] - 2020-06-15